update scripts/doc for mtls=true, refactorying sript structure, support custom cert

+++ demo/kserve/scripts/uninstall/kserve-uninstall.sh
@@ -27,10 +27,10 @@ export TARGET_OPERATOR_NS=$(getOpNS ${TARGET_OPERATOR_TYPE})
 
 oc delete validatingwebhookconfiguration inferencegraph.serving.kserve.io  inferenceservice.serving.kserve.io 
 oc delete mutatingwebhookconfiguration inferenceservice.serving.kserve.io
-oc delete isvc --all -n ${TEST_NS} --force --grace-period=0
+oc delete isvc,pod --all -n ${TEST_NS} --force --grace-period=0
 
 echo "It would take around around 3~4 mins"
-oc delete ns ${TEST_NS} ${MINIO_NS}
+oc delete ns ${TEST_NS} ${MINIO_NS} --force --grace-period=0
 oc delete secret wildcard-certs -n istio-system
 
 oc delete DataScienceCluster --all -n "${KSERVE_OPERATOR_NS}"
@@ -42,28 +42,15 @@ then
   oc delete catalogsource rhods-catalog-dev -n openshift-marketplace
 fi
 if [[ ${TARGET_OPERATOR_TYPE} == "rhods" ]]; then
-  oc delete ns redhat-ods-operator redhat-ods-applications rhods-notebooks redhat-ods-monitoring --force --grace-period=0
+  if [[ -n ${CLEAN_NS+x} && ${CLEAN_NS} != "false" ]]
+  then
+    oc delete ns redhat-ods-operator redhat-ods-applications rhods-notebooks redhat-ods-monitoring --force --grace-period=0
+  fi
   oc delete csv -n ${TARGET_OPERATOR_NS} $(oc get csv -n ${TARGET_OPERATOR_NS}|grep rhods|awk '{print $1}')
 else
-  oc delete ns ${KSERVE_OPERATOR_NS} --force --grace-period=0
+  if [[ -n ${CLEAN_NS+x} && ${CLEAN_NS} != "false" ]]
+  then  
+    oc delete ns ${KSERVE_OPERATOR_NS} --force --grace-period=0
+  fi
   oc delete csv -n ${TARGET_OPERATOR_NS} $(oc get csv -n ${TARGET_OPERATOR_NS} |grep opendatahub|awk '{print $1}')
 fi
-
-# Remove test namespace from SMMR
-INDEX=$(oc get servicemeshmemberroll/default -n istio-system -o jsonpath='{.spec.members[*]}')
-INDEX=$(echo ${INDEX} | tr ' ' '\n' | grep -n ${TEST_NS} | cut -d: -f1)
-
-if [ -z "${INDEX}" ]; then
-  echo "Target member ${TEST_NS} not found in the array."
-fi
-oc patch servicemeshmemberroll/default -n istio-system --type='json' -p="[{'op': 'remove', 'path': \"/spec/members/$((INDEX - 1))\"}]"
-
-
-# Remove kserve namespace from SMMR
-INDEX=$(oc get servicemeshmemberroll/default -n istio-system -o jsonpath='{.spec.members[*]}')
-INDEX=$(echo ${INDEX} | tr ' ' '\n' | grep -n ${KSERVE_OPERATOR_NS} | cut -d: -f1)
-
-if [ -z "${INDEX}" ]; then
-  echo "Target member ${TEST_NS} not found in the array."
-fi
-oc patch servicemeshmemberroll/default -n istio-system --type='json' -p="[{'op': 'remove', 'path': \"/spec/members/$((INDEX - 1))\"}]"
